blockstoragedevice: switch parameter for set_fpos to uint64_t.

+++ devices/storage/blockstoragedevice.h
@@ -39,7 +39,7 @@ public:
     int set_host_file(std::string file_path);
     int set_block_size(const int blk_size);
 
-    int set_fpos(const uint32_t lba);
+    int set_fpos(const uint64_t lba);
     int read_begin(int nblocks, uint32_t max_len = UINT32_MAX);
     int data_left() { return this->remain_size; }
     int read_more();
